refactor: modify interfaec of Print
