+++ printer/interface.go
@@ -1,9 +1,7 @@
 package printer
 
-import "bytes"
-
 // Interface qui abstrait l'impression
 type Printer interface {
-	Print(buffer bytes.Buffer) error
+	Print(content string) error
 	TestPrint() error
 }
