+++ printer/mock-printer.go
@@ -1,17 +1,15 @@
 package printer
 
 import (
-	"bytes"
-
 	log "github.com/sirupsen/logrus"
 )
 
 type MockPrinter struct{}
 
 // Implémentation pour afficher le texte dans la console (au lieu d'imprimer)
-func (p *MockPrinter) Print(buffer bytes.Buffer) error {
+func (p *MockPrinter) Print(content string) error {
 	log.Println("🔹 MOCK PRINTER OUTPUT 🔹")
-	log.Println(buffer.String())
+	log.Println(content)
 	log.Println("🔹 FIN 🔹")
 	return nil
 }
